Separate 'fit' functionality (new and old) into a mixin
